requirements.txt
streamlit
python-dotenv
langchain
langchain-chroma
langchain-google-genai
google-generativeai
chromadb
html5lib
google-genai
pysqlite3-binary
PyPDF2